Add/azure reverse proxy/listener

+++ AzureReverseProxy/AzureReverseProxy/AzureReverseProxyCMD/Program.cs
@@ -1,5 +1,9 @@
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
 using System;
 using System.IO;
+using System.Threading.Tasks;
 using AzureReverseProxy;
 using Newtonsoft.Json;
 
@@ -8,15 +12,46 @@ namespace AzureReverseProxyCMD
     class Program
     {
         private static AzureReverseTool AZProxy;
+        private static bool DeleteResource = false;
+        private static bool Run = true;
         static void Main(string[] args)
         {
             AppDomain.CurrentDomain.ProcessExit += new EventHandler(OnProcessExit);
 
-            var config = GetConfiguration();
+            // Event to stop the process
+            Console.CancelKeyPress += delegate (object sender, ConsoleCancelEventArgs e) {
+                e.Cancel = true;
+                Run = false;
+                AZProxy.CloseListener().GetAwaiter().GetResult();
+            };
 
-            Console.WriteLine($"Deploying { config.DeploymentName } to group { config.ResourceGroupName }...");
+            MainAsync().Wait();
+        }
+
+        private static async Task MainAsync()
+        {
+            var config = GetConfiguration();
             AZProxy = new AzureReverseTool(config);
-            AZProxy.GenerateRelayResource();
+
+            Console.WriteLine($"Check existence of '{ config.DeploymentName }' in '{ config.ResourceGroupName }' group...");
+            if (!AZProxy.CheckExistence())
+            {
+                Console.WriteLine($"'{ config.DeploymentName }' was not found. Deploying to '{ config.ResourceGroupName }' group...");
+                AZProxy.GenerateRelayResource();
+                Console.WriteLine($"'{ config.DeploymentName }' created.");
+            }
+            else
+            {
+                Console.WriteLine($"'{ config.DeploymentName }' found.");
+            }
+
+            Console.WriteLine("Starting listener...\n");
+            await AZProxy.StartListener();
+
+            while (Run)
+            {
+                await Task.Delay(100);
+            }
         }
 
         private static IConfiguration GetConfiguration()
@@ -32,8 +67,17 @@ namespace AzureReverseProxyCMD
 
         private static void OnProcessExit(object sender, EventArgs e)
         {
-            Console.WriteLine(Environment.NewLine + "Deleting the generated resources...");
-            AZProxy.DeleteResource();
+            if (AZProxy != null)
+            {
+                AZProxy.CloseListener().GetAwaiter().GetResult();
+            }
+
+            if (DeleteResource)
+            {
+                Console.WriteLine(Environment.NewLine + "Deleting the generated resources...");
+                AZProxy.DeleteResource();
+                Console.WriteLine("\nThe resurce was eliminated.");
+            }
         }
     }
 }
